% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{cluster.results}
\alias{cluster.results}
\title{Function to construct similarity matrix based on the response rates}
\usage{
cluster.results(
  group,
  n,
  r,
  alpha = 1,
  kmin = 2,
  kmax = 4,
  lambda = 10^c(5, 10, 20, 40, 80),
  mu0 = 0,
  sigma0 = 10,
  sigma_y = 0.001,
  burnIn = 1000,
  maxIters = 1000,
  ...
)
}
\arguments{
\item{group}{A vector of subgroup labels.}

\item{n}{A vector representing the number of patients in each subgroup.}

\item{r}{A vector representing the number of responses in each subgroup.}

\item{alpha}{Numeric. The concentration parameter of the Dirichlet Process.}

\item{kmin}{Integer. The minimum number of components (K) to consider in the Bayesian Finite Gaussian Mixture Model (BFGMM).}

\item{kmax}{Integer. The maximum number of components (K) to consider in the BFGMM.}

\item{lambda}{A numeric vector of tuning parameters for the modified Chinese restaurant process mixed model (m-CRPMM).}

\item{mu0}{Numeric. The hyperprior mean for the BFGMM and m-CRPMM models.}

\item{sigma0}{Numeric. The hyperprior variance for the BFGMM and m-CRPMM models.}

\item{sigma_y}{Numeric. Data measurement error for the BFGMM and m-CRPMM models.}

\item{burnIn}{Integer. Number of burn-in iterations for Gibbs sampling.}

\item{maxIters}{Integer. The number of MCMC samples.}
}
\value{
A list containing the following components:
  \item{data}{The original dataset of response rates provided to the function.}
  \item{summary}{A summary of the clustering analysis}
  \item{optmethod}{The optimal clustering method used.}
  \item{opttab}{MCMC samples from the optimal clustering method used.}
  \item{sm}{Similarity matrix based on clustering results.}
}
\description{
Function to construct similarity matrix based on the response rates
}
