% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{esg_filter}
\alias{esg_filter}
\title{Function to filter extremely small groups}
\usage{
esg_filter(
  group,
  n,
  r,
  adj_p = 0.5,
  aprior = 0.5,
  bprior = 0.5,
  eps = 0.15,
  minPts = 4,
  iter = 1000,
  seeds = 7150321,
  ...
)
}
\arguments{
\item{group}{A vector of subgroup labels.}

\item{n}{A vector representing the number of patients in each subgroup.}

\item{r}{A vector representing the number of responses in each subgroup.}

\item{adj_p}{Numeric. Adjustment parameter for adjusting responses.}

\item{aprior}{Numeric. Prior parameter 'a' for Bayesian estimation.}

\item{bprior}{Numeric. Prior parameter 'b' for Bayesian estimation.}

\item{eps}{Numeric. The epsilon parameter for the DBSCAN clustering algorithm.}

\item{minPts}{Integer. Minimum number of points to form a dense region in DBSCAN.}

\item{iter}{Number of iterations for the filtering process.}

\item{seeds}{Integer. Random seed for reproducibility.}
}
\value{
A data frame containing the original group information with an added column
        `esg.prob`, which gives the probability of each group being identified as a
        extremely small group across all iterations.
}
\description{
This function filters small groups based on a perturbation analysis
and Jensen-Shannon Divergence (JSD) using the DBSCAN clustering algorithm.
}
