% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{GF.samp}
\alias{GF.samp}
\title{Perform Gibbs Sampling for a Finite Gaussian Mixture Model}
\usage{
GF.samp(
  data,
  k,
  burnIn = 1000,
  maxIters = 1000,
  mu0 = 0,
  sigma0 = 10,
  sigma_y = 0.001,
  a = 10000
)
}
\arguments{
\item{data}{A numeric vector or matrix representing the observed data points.}

\item{k}{Integer. Pre-specified the number of clusters.}

\item{burnIn}{Integer. Number of burn-in iterations for Gibbs sampling.}

\item{maxIters}{Integer. The number of MCMC samples.}

\item{mu0}{Numeric. Prior mean for the Gaussian distribution of cluster means.}

\item{sigma0}{Numeric. Prior variance for the Gaussian distribution of cluster means.}

\item{sigma_y}{Numeric. Measurement error variance.}

\item{a}{Numeric. Dirichlet prior parameters.}
}
\value{
A matrix where each row corresponds to the cluster assignments of the data points at a given iteration (after burn-in).
}
\description{
This function performs Gibbs sampling for a finite Gaussian mixture model.
}
