% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{GI.samp}
\alias{GI.samp}
\title{Perform Gibbs Sampling for a Gaussian infinite mixture model}
\usage{
GI.samp(
  data,
  lambda = 1,
  alpha = 1e-20,
  mu0 = 0,
  sigma0 = 10,
  sigma_y = 0.001,
  burnIn = 1000,
  maxIters = 1000
)
}
\arguments{
\item{data}{A numeric vector (or n x 1 matrix) of data points to cluster.}

\item{lambda}{Numeric. Tuning parameters for the modified Chinese restaurant process mixed model.}

\item{alpha}{Numeric. The concentration parameter of the Dirichlet Process.}

\item{mu0}{Numeric. Prior mean for the Gaussian distribution of cluster means.}

\item{sigma0}{Numeric. Prior variance for the Gaussian distribution of cluster means.}

\item{sigma_y}{Numeric. Measurement error variance.}

\item{burnIn}{Integer. Number of burn-in iterations for Gibbs sampling.}

\item{maxIters}{Integer. The number of MCMC samples.}
}
\value{
A matrix where each row corresponds to the cluster assignments of the data points at a given iteration (after burn-in).
}
\description{
This function performs Gibbs sampling on a Gaussian infinite mixture model
using a modified Chinese Restaurant Process (CRP) prior. It generates cluster
assignments for the data based on the Dirichlet Process.
}
