% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{logit}
\alias{logit}
\title{Logit Transformation
This function computes the logit transformation of a given numeric input.}
\usage{
logit(x)
}
\description{
Logit Transformation
This function computes the logit transformation of a given numeric input.
}
