% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BACH.R
\name{BACH}
\alias{BACH}
\title{BACH Function for Analyzing Concurrent Response Rate Data with Heterogeneity in Basket Trials}
\usage{
BACH(
  group,
  n,
  r,
  esfilter = T,
  escutoff = 0.8,
  kmin = 2,
  kmax = 4,
  lambda = 10^c(5, 10, 20, 40, 80),
  mu0 = 0.2,
  sigma0 = 10,
  sigma_y = 0.001,
  w = 0.95,
  alpha = 1,
  alpha1 = 5,
  beta1 = 1,
  tau2 = 0.1,
  c.burnin = 1000,
  c.maxiters = 1000,
  b.burnin = 5000,
  b.maxiters = 10000,
  seeds = 2024
)
}
\arguments{
\item{group}{A vector of subgroup labels.}

\item{n}{A vector representing the number of patients in each subgroup.}

\item{r}{A vector representing the number of responses in each subgroup.}

\item{esfilter}{Logical. Whether to filter out extremely small subgroups. Default is TRUE.}

\item{escutoff}{Numeric. The cutoff probability used to identify extremely small subgroups.}

\item{kmin}{Integer. The minimum number of components (K) to consider in the Bayesian Finite Gaussian Mixture Model (FGMM).}

\item{kmax}{Integer. The maximum number of components (K) to consider in the BFGMM.}

\item{lambda}{A numeric vector of tuning parameters for Bayesian Infinite Gaussian Mixture Model (IGMM).}

\item{mu0}{Numeric. The hyperprior mean for the FGMM and IGMM models.}

\item{sigma0}{Numeric. The hyperprior variance for the FGMM and IGMM models.}

\item{sigma_y}{Numeric. Data measurement error for the FGMM and IGMM models.}

\item{w}{Numeric. A weight parameter (between 0 and 1) that determines the relative importance of subgroup similarity matrices \code{Cij} and \code{Dij}.}

\item{alpha}{Numeric. The concentration parameter (alpha) for the Dirichlet Process in IGMM.}

\item{alpha1}{Numeric. Gamma Prior shape parameter in the modified Bayesian Hierarchical Model (BHM).}

\item{beta1}{Numeric. Gamma Prior scale parameter in the modified BHM.}

\item{tau2}{Numeric. The hyperprior precision parameter for subgroup means in the modified BHM.}

\item{c.burnin}{Integer. The number of burn-in samples for the FGMM and IGMM models.}

\item{c.maxiters}{Integer. The number of MCMC iterations for the FGMM and IGMM models.}

\item{b.burnin}{Integer. The number of burn-in samples for the BHM model. Default is 5000.}

\item{b.maxiters}{Integer. The number of MCMC iterations for the BHM model. Default is 10000.}

\item{seeds}{Integer. Random seed for reproducibility.}
}
\value{
A list containing the following components:
  \item{optmethod}{The optimal clustering method used.}
  \item{mc}{Similarity matrix based on clustering results.}
  \item{mw}{Weighted similarity matrix combining clustering results and subgroup size similarities.}
  \item{Summary}{A data frame summarizing the posterior mean, 95\% credible interval, and other information for each subgroup.}
}
\description{
BACH Function for Analyzing Concurrent Response Rate Data with Heterogeneity in Basket Trials
}
